+++ src/Billing/Billing/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Removed "Microsoft.Azure.Management.Billing" Version "4.1.0" PackageReference
+* Removed "Microsoft.Azure.Management.Consumption" Version "3.0.0" PackageReference
 * Upgraded nuget package to signed package.
 
 ## Version 2.1.0